Split out module rewriting for CollapseProperties to its own test class
