+++ src/com/google/javascript/jscomp/CollapseProperties.java
@@ -253,6 +253,9 @@ class CollapseProperties implements CompilerPass {
     }
 
     for (Name name : nameMap.values()) {
+      if (dynamicallyImportedModuleRefs.contains(name.getFullName())) {
+        escaped.add(name);
+      }
       if (!name.isNamespaceObjectLit()) {
         continue;
       }
